+++ packages/odyssey-react-mui/src/Card.tsx
@@ -24,10 +24,8 @@ import {
 } from "@mui/material";
 import styled from "@emotion/styled";
 
-import { Button } from "./Button";
-import { ButtonContext } from "./ButtonContext";
+import { Button, ButtonContext, MenuButton, MenuButtonProps } from "./Buttons";
 import { MoreIcon } from "./icons.generated";
-import { MenuButton, MenuButtonProps } from "./MenuButton";
 import {
   DesignTokens,
   useOdysseyDesignTokens,