+++ packages/odyssey-react-mui/src/DataTable/DataTableRowActions.tsx
@@ -12,10 +12,8 @@
 
 import { MRT_Row, MRT_RowData } from "material-react-table";
 import { Fragment, ReactElement, memo, useCallback } from "react";
-import { Button } from "../Button";
-import { MenuItem } from "../MenuItem";
+import { Button, MenuButton, MenuButtonProps, MenuItem } from "../Buttons";
 import { Box as MuiBox } from "@mui/material";
-import { MenuButton, MenuButtonProps } from "../MenuButton";
 import {
   ArrowBottomIcon,
   ArrowDownIcon,