+++ packages/odyssey-react-mui/src/labs/AppTile.tsx
@@ -25,7 +25,7 @@ import {
 } from "@mui/material";
 import styled from "@emotion/styled";
 
-import { Button } from ".././Button";
+import { Button } from "../Buttons";
 import {
   DesignTokens,
   useOdysseyDesignTokens,