+++ packages/odyssey-react-mui/src/labs/DataView/DataCard.tsx
@@ -29,14 +29,17 @@ import styled from "@emotion/styled";
 import { useTranslation } from "react-i18next";
 
 import { Box } from "../../Box";
-import { Button } from "../../Button";
-import { ButtonContext } from "../../ButtonContext";
+import {
+  Button,
+  ButtonContext,
+  MenuButton,
+  MenuButtonProps,
+} from "../../Buttons";
 import {
   DesignTokens,
   useOdysseyDesignTokens,
 } from "../../OdysseyDesignTokensContext";
 import { Heading5, Paragraph, Support } from "../../Typography";
-import { MenuButton, MenuButtonProps } from "../../MenuButton";
 import {
   ChevronDownIcon,
   ChevronUpIcon,