+++ packages/odyssey-react-mui/src/labs/DataView/DataView.tsx
@@ -35,7 +35,7 @@ import { DataFilters } from "../DataFilters";
 import { EmptyState } from "../../EmptyState";
 import { fetchData } from "./fetchData";
 import { LayoutSwitcher } from "./LayoutSwitcher";
-import { MenuButton } from "../../MenuButton";
+import { MenuButton } from "../../Buttons";
 import { MoreIcon } from "../../icons.generated";
 import { TableSettings } from "./TableSettings";
 import { Pagination, usePagination } from "../../Pagination";
@@ -101,6 +101,9 @@ const DataView = ({
   getRowId: getRowIdProp,
   hasFilters,
   hasPagination,
+  hasPageInput,
+  hasRowCountInput,
+  hasRowCountLabel,
   hasSearch,
   hasSearchSubmitButton,
   hasRowReordering,
@@ -461,6 +464,9 @@ const DataView = ({
       {hasPagination && (
         <Pagination
           currentPageLabel={t("pagination.page")}
+          hasPageInput={hasPageInput}
+          hasRowCountInput={hasRowCountInput}
+          hasRowCountLabel={hasRowCountLabel}
           isDisabled={isEmpty}
           isMoreDisabled={isPaginationMoreDisabled}
           lastRow={lastRowOnPage}